src/main/java/ar/com/codoacodo/flysky/controller/ClasesController.java
package ar.com.codoacodo.flysky.controller;

import ar.com.codoacodo.flysky.model.dto.DtoClase;
import ar.com.codoacodo.flysky.service.ClasesServiceImpl;
import lombok.AllArgsConstructor;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@AllArgsConstructor
@RestController
@RequestMapping("/api/clases")
public class ClasesController {

    @Autowired
    private ClasesServiceImpl service;

    @GetMapping("/")
    public ResponseEntity<?> listarTodos() {
        return new ResponseEntity<>(service.todasLasClases(), HttpStatus.OK);
    }

    @GetMapping("/{id}")
    public ResponseEntity<?> seleccionar(@PathVariable Integer id) {
        return new ResponseEntity<>(service.buscarPorId(id), HttpStatus.OK);
    }

    @PostMapping("/registrar")
    public ResponseEntity<?> registrar(@RequestBody DtoClase clase) {
        return new ResponseEntity<>(service.agregarNueva(clase), HttpStatus.OK);
    }

    @DeleteMapping("/{id}/eliminar")
    public ResponseEntity<?> eliminar(@PathVariable Integer id) {
        return new ResponseEntity<>(service.eliminar(id), HttpStatus.OK);
    }

}
